implementations/feature1_account_management/requirements.txt
Flask
psycopg2-binary
Werkzeug
PyJWT